Made DreamBooster abstract, move actual Dream Booster to DreamBoosterSegment
