+++ src/Entities/BoosterStuff/DreamBooster.cs
@@ -9,74 +9,7 @@ using System;
 using System.Collections;
 
 namespace Celeste.Mod.CommunalHelper.Entities {
-    [CustomEntity("CommunalHelper/DreamBooster")]
-    public class DreamBooster : CustomBooster {
-
-        public class DreamBoosterPathRenderer : Entity {
-            private DreamBooster dreamBooster;
-
-            public float Alpha;
-            public float Percent = 1f;
-            public float RainbowLerp;
-
-            private Vector2 perp;
-
-            public DreamBoosterPathRenderer(DreamBooster booster, float alpha) {
-                Depth = Depths.DreamBlocks + 1;
-                dreamBooster = booster;
-
-                Alpha = Percent = alpha;
-                perp = dreamBooster.Dir.Perpendicular();
-            }
-
-            public override void Update() {
-                base.Update();
-                if (dreamBooster.BoostingPlayer)
-                    RainbowLerp += Engine.DeltaTime * 8f;
-            }
-
-            public override void Render() {
-                base.Render();
-                if (Alpha <= 0f)
-                    return;
-
-                Color color = dreamBooster.BoostingPlayer ? Util.ColorArrayLerp(RainbowLerp, DreamColors) : Color.White;
-
-                Util.TryGetPlayer(out Player player);
-                for (float f = 0f; f < dreamBooster.Length * Percent; f += 6f) {
-                    DrawPathLine(f, player, color);
-                }
-                DrawPathLine(dreamBooster.Length * Percent - dreamBooster.Length % 6, null, Color.White);
-            }
-
-            private void DrawPathLine(float linePos, Player player, Color lerp) {
-                Vector2 pos = dreamBooster.Start + dreamBooster.Dir * linePos;
-                float sin = (float) Math.Sin(linePos + Scene.TimeActive * 6f) * 0.3f + 1f;
-
-                float highlight = player == null ? 0.25f : Calc.ClampedMap(Vector2.Distance(player.Center, pos), 0, 80);
-                float lineHighlight = (1 - highlight) * 2.5f + 0.75f;
-                float alphaHighlight = 1 - Calc.Clamp(highlight, 0.01f, 0.8f);
-                Color color = Color.Lerp(Color.White, lerp, 1 - highlight) * alphaHighlight;
-
-                float lineLength = lineHighlight * sin;
-                Vector2 lineOffset = perp * lineLength;
-
-                // Single perpendicular short segments
-                //Draw.Line(pos + lineOffset, pos - lineOffset, color * Alpha);
-
-                // "Arrow" style
-                Vector2 arrowOffset = -dreamBooster.Dir * lineLength;
-                Draw.Line(pos, pos - lineOffset + arrowOffset, color * Alpha);
-                Draw.Line(pos, pos + lineOffset + arrowOffset, color * Alpha);
-            }
-        }
-
-        private DreamBoosterPathRenderer pathRenderer;
-        private bool showPath = true;
-
-        public float Length;
-        public Vector2 Start, Target, Dir;
-
+    public abstract class DreamBooster : CustomBooster {
         public static ParticleType P_BurstExplode;
 
         public static readonly Color BurstColor = Calc.HexToColor("19233b");
@@ -95,20 +28,8 @@ namespace Celeste.Mod.CommunalHelper.Entities {
         };
         public static readonly ParticleType[] DreamParticles = new ParticleType[8];
 
-        public DreamBooster(EntityData data, Vector2 offset)
-            : this(data.Position + offset, data.Nodes[0] + offset, !data.Bool("hidePath")) { }
-
-        public DreamBooster(Vector2 position, Vector2 node, bool showPath)
-            : base(position, redBoost: true) {
-            Depth = Depths.DreamBlocks;
-
-            Target = node;
-            Dir = Calc.SafeNormalize(Target - Position);
-            Length = Vector2.Distance(position, Target);
-            Start = position;
-
-            this.showPath = showPath;
-
+        public DreamBooster(Vector2 position, bool showPath)
+            : base (position, redBoost: true) {
             ReplaceSprite(CommunalHelperModule.SpriteBank.Create("dreamBooster"));
             SetParticleColors(BurstColor, AppearColor);
             SetSoundEvent(
@@ -117,52 +38,7 @@ namespace Celeste.Mod.CommunalHelper.Entities {
                 false);
         }
 
-        public override void Added(Scene scene) {
-            base.Added(scene);
-            scene.Add(pathRenderer = new DreamBoosterPathRenderer(this, Util.ToInt(showPath)));
-        }
-
-        public override void Removed(Scene scene) {
-            base.Removed(scene);
-            scene.Remove(pathRenderer);
-            pathRenderer = null;
-            base.Removed(scene);
-        }
-
-        protected override void OnPlayerEnter(Player player) {
-            base.OnPlayerEnter(player);
-            pathRenderer.RainbowLerp = Calc.Random.Range(0, 8);
-            if (!showPath)
-                Add(new Coroutine(HiddenPathReact()));
-        }
-
-        private IEnumerator HiddenPathReact() {
-            float duration = 0.5f;
-            float timer = 0f;
-
-            showPath = true;
-            SetSoundEvent(
-                CustomSFX.game_customBoosters_dreamBooster_dreambooster_enter,
-                CustomSFX.game_customBoosters_dreamBooster_dreambooster_move,
-                false);
-
-            ParticleSystem particlesBG = SceneAs<Level>().ParticlesBG;
-            while (timer < duration) {
-                timer += Engine.DeltaTime;
-                pathRenderer.Alpha = pathRenderer.Percent = Ease.SineOut(timer / duration);
-
-                Vector2 pos = Start + Dir * pathRenderer.Percent * Length;
-
-                particlesBG.Emit(DreamParticles[Calc.Random.Range(0, 8)], 2, pos, Vector2.One * 2f, (-Dir).Angle());
-                yield return null;
-            }
-        }
-
-        public override void Render() {
-            base.Render();
-        }
-
-        public static void InitializeParticles() {
+        internal static void InitializeParticles() {
             P_BurstExplode = new ParticleType(P_Burst) {
                 Color = BurstColor,
                 SpeedMax = 250
@@ -205,7 +81,7 @@ namespace Celeste.Mod.CommunalHelper.Entities {
                 yield return origEnum.Current;
 
             // could have done this in Booster.PlayerReleased, but it doesn't pass the player object
-            if (self is DreamBooster booster) {
+            if (self is DreamBoosterSegment booster) {
                 float angle = booster.Dir.Angle() - 0.5f;
                 for (int i = 0; i < 20; i++) {
                     booster.SceneAs<Level>().ParticlesBG.Emit(DreamBooster.P_BurstExplode, 1, player.Center, new Vector2(3f, 3f), angle + Calc.Random.NextFloat());
@@ -214,9 +90,9 @@ namespace Celeste.Mod.CommunalHelper.Entities {
         }
 
         private static int Player_RedDashUpdate(On.Celeste.Player.orig_RedDashUpdate orig, Player self) {
-            DreamBooster dreamBooster = self.LastBooster as DreamBooster;
+            DreamBoosterSegment dreamBooster = self.LastBooster as DreamBoosterSegment;
             if (dreamBooster != null) {
-                bool inSolid = self.CollideCheck<Solid, DreamBlock>();
+                bool inSolid = self.CollideCheck<Solid>();
 
                 // Prevent the player from jumping or dashing out of the DreamBooster. May be reset in IL hook below.
                 // If for whatever reason this becomes an actual option for DreamBoosters, this will need to be changed.
@@ -260,7 +136,7 @@ namespace Celeste.Mod.CommunalHelper.Entities {
             while (origEnum.MoveNext())
                 yield return origEnum.Current;
 
-            if (currentBooster is DreamBooster booster) {
+            if (currentBooster is DreamBoosterSegment booster) {
                 DynData<Player> playerData = new DynData<Player>(self);
                 self.Speed = ((Vector2) (playerData["gliderBoostDir"] = self.DashDir = booster.Dir)) * 240f;
                 self.SceneAs<Level>().DirectionalShake(self.DashDir, 0.2f);
